requirements.txt
python-telegram-bot[job-queue]
openai
google-auth
google-auth-oauthlib
google-api-python-client
sqlalchemy
python-dotenv
pydantic
pydantic-settings
pytz
aiosqlite
pytest
pytest-asyncio
google-generativeai